refactor: ドメインモデルの定義と型の再構築（P1）
- RunState を 9 つのサブステート型に分割（PlayerState, BattleState, ProgressState 等）
- types/ ディレクトリに 18 ファイルの型定義を新設
- RunState を extends による合成型として再定義し後方互換を維持
- GamePhase ステートマシン（PHASE_TRANSITIONS, assertValidTransition）を追加
- types.ts を barrel re-export に変換
- TDD で 20 件のテストを追加（型検証 + フェーズ遷移テスト）
- 全 229 スイート / 3209 テストパス、型チェック・ビルド成功確認済み
